Support multiple games via new state objects
Clients `clients` is the new private state, just an array of data for
each client containing nickname, gameId, token and key (websocket id).

Games `games` is the new publicState scoped on gameId, for each game
there is maze (still double dimension array) and players (nickname, x,
y, direction, speed).

Player speed starts out as 0 (stopped), and becomes 1 (driving) when
driving. (There is no way of getting it to 0 again yet)

We now have incremental push of state updates via the setInterval and
incrementalStorage.

Every tick we take the queue of changes done for a game and push them
out to each player. For example:

[['game123', 'players', 'albert', 'direction', 'NORTH'
 ['game123', 'maze', 4, 7, 'NORTH']
 ['game123', 'players', 'kevin', 'x', 10]]

We now also have an initial direction for players based of where the
entrance is.

🎉